cpufreq: interactive: Skip timer when policy->min == policy->max
When the minfreq is set to the maxfreq, there is nothing to really do: the
CPU is already running at its maxfreq. In this scenario, just do nothing
and re-arm the timer.

This scenario occurs when the screen is turned on due to the cpu_input_boost
driver (it sets policy->min to policy->max for a few seconds).

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: imzf <tdzfeng@gmail.com>
Signed-off-by: Ahmad Thoriq Najahi <tlogitechnjhi@gmail.com>
Signed-off-by: keizzn <keiusuke.faisal123@gmail.com>
Signed-off-by: FauzanDhiyaul <fdhiyaul163@gmail.com>